satisfy clippy warnings

+++ src/utility/rgb.rs
@@ -28,9 +28,9 @@ pub fn write_pixel(
     g = f32::sqrt(scale * g);
     b = f32::sqrt(scale * b);
 
-    let ir = (256 as f32 * clamp(r, 0.0, 0.999)) as u8;
-    let ig = (256 as f32 * clamp(g, 0.0, 0.999)) as u8;
-    let ib = (256 as f32 * clamp(b, 0.0, 0.999)) as u8;
+    let ir = (256_f32 * clamp(r, 0.0, 0.999)) as u8;
+    let ig = (256_f32 * clamp(g, 0.0, 0.999)) as u8;
+    let ib = (256_f32 * clamp(b, 0.0, 0.999)) as u8;
 
     buffer.put_pixel(x as u32, y as u32, image::Rgb([ir, ig, ib]));
 }
